Removed unnecessary idling from tests.
