--- sdk/src/androidTest/java/com/desk/android/sdk/util/ElapsedTimeIdlingResource.java
@@ -1,41 +1,0 @@
-package com.desk.android.sdk.util;
-
-import android.support.test.espresso.IdlingResource;
-
-/**
- * TODO - Add class comments
- * <p/>
- * Created by Jerrell Mardis
- * Copyright (c) 2015 Desk.com. All rights reserved.
- */
-public class ElapsedTimeIdlingResource implements IdlingResource {
-
-    private final long startTime;
-    private final long waitingTime;
-    private ResourceCallback resourceCallback;
-
-    public ElapsedTimeIdlingResource(long waitingTime) {
-        this.startTime = System.currentTimeMillis();
-        this.waitingTime = waitingTime;
-    }
-
-    @Override
-    public String getName() {
-        return ElapsedTimeIdlingResource.class.getName() + ":" + waitingTime;
-    }
-
-    @Override
-    public boolean isIdleNow() {
-        long elapsed = System.currentTimeMillis() - startTime;
-        boolean idle = (elapsed >= waitingTime);
-        if (idle) {
-            resourceCallback.onTransitionToIdle();
-        }
-        return idle;
-    }
-
-    @Override
-    public void registerIdleTransitionCallback(ResourceCallback resourceCallback) {
-        this.resourceCallback = resourceCallback;
-    }
-}
