+++ sdk/src/androidTest/java/com/desk/android/sdk/util/InstrumentationTestUtils.java
@@ -2,11 +2,7 @@ package com.desk.android.sdk.util;
 
 import android.support.annotation.NonNull;
 import android.support.test.InstrumentationRegistry;
-import android.support.test.espresso.Espresso;
-import android.support.test.espresso.IdlingPolicies;
-import android.support.test.espresso.IdlingResource;
 import android.support.test.espresso.ViewInteraction;
-import android.support.test.espresso.matcher.BoundedMatcher;
 import android.support.v7.widget.Toolbar;
 
 import com.desk.android.sdk.Desk;
@@ -17,14 +13,10 @@ import com.desk.java.apiclient.model.Article;
 import com.desk.java.apiclient.model.Topic;
 import com.google.gson.reflect.TypeToken;
 
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 import static com.desk.android.sdk.util.TestUtils.readMockJsonFile;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.hamcrest.core.Is.is;
 
 /**
@@ -36,19 +28,18 @@ import static org.hamcrest.core.Is.is;
 public class InstrumentationTestUtils {
 
     public static ApiResponse<Article> getMockedArticleResponse() {
-        ApiResponse<Article> response = readMockJsonFile(
+        return readMockJsonFile(
                 new TypeToken<ApiResponse<Article>>() {}.getType(),
                 "mock/mock_article_response.json"
         );
-        return response;
     }
 
     public static ApiResponse<Topic> getMockedTopicResponse() {
-        ApiResponse<Topic> response = readMockJsonFile(
-                new TypeToken<ApiResponse<Topic>>() {}.getType(),
+        return readMockJsonFile(
+                new TypeToken<ApiResponse<Topic>>() {
+                }.getType(),
                 "mock/mock_topic_response.json"
         );
-        return response;
     }
 
     @NonNull
@@ -58,35 +49,8 @@ public class InstrumentationTestUtils {
         return isWebFormEnabled ? ContactUsWebActivity.class.getName() : ContactUsActivity.class.getName();
     }
 
-    @NonNull
-    public static IdlingResource createIdleResourceAndWait(long waitingTime) {
-        // set Espresso timeout policies
-        IdlingPolicies.setMasterPolicyTimeout(waitingTime * 2, MILLISECONDS);
-        IdlingPolicies.setIdlingResourceTimeout(waitingTime * 2, MILLISECONDS);
-
-        // wait
-        IdlingResource idlingResource = new ElapsedTimeIdlingResource(waitingTime);
-        Espresso.registerIdlingResources(idlingResource);
-        return idlingResource;
-    }
-
     public static ViewInteraction matchToolbarTitle(CharSequence title) {
         return onView(isAssignableFrom(Toolbar.class))
-                .check(matches(withToolbarTitle(is(title))));
-    }
-
-    public static Matcher<Object> withToolbarTitle(final Matcher<CharSequence> textMatcher) {
-        return new BoundedMatcher<Object, Toolbar>(Toolbar.class) {
-            @Override
-            public boolean matchesSafely(Toolbar toolbar) {
-                return textMatcher.matches(toolbar.getTitle());
-            }
-
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("with toolbar title: ");
-                textMatcher.describeTo(description);
-            }
-        };
+                .check(matches(ViewMatchers.withToolbarTitle(is(title))));
     }
 }
